variant: new internal flag to mark referenced strings or bytes
When creating variants as strings or bytes, upon creation the bytes for those
data types are allocated in a new memory buffer, it's a copy of what the caller
passed as a reference. There are cases where due to performance and
optimization reasons the caller might want to keep a reference to the original
buffer rather than a full copy.

This patch adds a new flag called 'referenced' that can be used by the new API
changes coming in the new series of patches.

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>