variant: refactor string and bytes handling API
The following patch does prototype changes for functions that create variants of
string and byte types, as well it adds two new functions to set the size/lengths
of the buffers. The new prototypes are as follows:

 - struct cfl_variant *cfl_variant_create_from_string_s(char *value, size_t value_size, int referenced)
 - struct cfl_variant *cfl_variant_create_from_bytes(char *value, size_t length, int referenced)

the new `referenced` flag marks if the caller desires to keep a reference to the original buffer
rather a full copy in a new buffer.

Also this patch adds two new functions to set and retrieve the size of such buffers:

- void cfl_variant_size_set(struct cfl_variant *var, size_t size);
- size_t cfl_variant_size_get(struct cfl_variant *var);

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>